+++ R/svnLog.R
@@ -1,15 +1,19 @@
 #' Get the SVN history of a file
 #' 
 #' @description 
-#' Returns the raw output of `svn log` and `svn info` for the specified
-#' file. `svn log` provides information regarding each revision of the 
-#' file and `svn info` provides an overview of the latest update.
+#' Returns the `svn log` in dataframe format.
 #' 
-#' @param .file Relative file path of the file of interest
+#' @param .file File path of the file of interest
 #' 
 #' @export
 svnLog <- function(.file) {
   
-  logdf <- svnCommand(.file = .file, .command = "log")
-  dplyr::bind_rows(logdf) %>% dplyr::rename(rev = .attrs)
+  log_df <- svnCommand(.file = .file, .command = "log")
+  log_return <- 
+    dplyr::bind_rows(log_df) %>% 
+    dplyr::rename(rev = .attrs) %>% 
+    dplyr::mutate(datetime = as.POSIXct(date, format="%Y-%m-%dT%H:%M:%OS", tz="UTC")) %>% 
+    dplyr::select(author, datetime, rev, msg)
+  
+  return(log_return)
 }